+++ benches/bench.rs
@@ -1,5 +1,5 @@
 // This benchmark suite contains some benchmarks along a set of dimensions:
-//   Hasher: std default (SipHash) and crate default (foldhash).
+//   Hasher: std default (SipHash) and crate default (rapidhash).
 //   Int key distribution: low bit heavy, top bit heavy, and random.
 //   Task: basic functionality: insert, insert_erase, lookup, lookup_fail, iter
 #![feature(test)]
@@ -19,7 +19,7 @@ const SIZE: usize = 1000;
 const OP_COUNT: usize = 500;
 
 // The default hashmap when using this crate directly.
-type FoldHashMap<K, V> = HashMap<K, V, DefaultHashBuilder>;
+type RapidHashMap<K, V> = HashMap<K, V, DefaultHashBuilder>;
 // This uses the hashmap from this crate with the default hasher of the stdlib.
 type StdHashMap<K, V> = HashMap<K, V, RandomState>;
 
@@ -57,14 +57,14 @@ impl Drop for DropType {
 }
 
 macro_rules! bench_suite {
-    ($bench_macro:ident, $bench_foldhash_serial:ident, $bench_std_serial:ident,
-     $bench_foldhash_highbits:ident, $bench_std_highbits:ident,
-     $bench_foldhash_random:ident, $bench_std_random:ident) => {
-        $bench_macro!($bench_foldhash_serial, FoldHashMap, 0..);
+    ($bench_macro:ident, $bench_rapidhash_serial:ident, $bench_std_serial:ident,
+     $bench_rapidhash_highbits:ident, $bench_std_highbits:ident,
+     $bench_rapidhash_random:ident, $bench_std_random:ident) => {
+        $bench_macro!($bench_rapidhash_serial, RapidHashMap, 0..);
         $bench_macro!($bench_std_serial, StdHashMap, 0..);
         $bench_macro!(
-            $bench_foldhash_highbits,
-            FoldHashMap,
+            $bench_rapidhash_highbits,
+            RapidHashMap,
             (0..).map(usize::swap_bytes)
         );
         $bench_macro!(
@@ -72,7 +72,7 @@ macro_rules! bench_suite {
             StdHashMap,
             (0..).map(usize::swap_bytes)
         );
-        $bench_macro!($bench_foldhash_random, FoldHashMap, RandomKeys::new());
+        $bench_macro!($bench_rapidhash_random, RapidHashMap, RandomKeys::new());
         $bench_macro!($bench_std_random, StdHashMap, RandomKeys::new());
     };
 }
@@ -82,14 +82,14 @@ macro_rules! bench_suite_2 {
      $name0:ident, $size0:literal, $name1:ident, $size1:literal, $name2:ident, $size2:literal,
      $name3:ident, $size3:literal, $name4:ident, $size4:literal, $name5:ident, $size5:literal,
      $name6:ident, $size6:literal, $name7:ident, $size7:literal) => {
-        $bench_macro!($name0, $size0, FoldHashMap, RandomKeys::new());
-        $bench_macro!($name1, $size1, FoldHashMap, RandomKeys::new());
-        $bench_macro!($name2, $size2, FoldHashMap, RandomKeys::new());
-        $bench_macro!($name3, $size3, FoldHashMap, RandomKeys::new());
-        $bench_macro!($name4, $size4, FoldHashMap, RandomKeys::new());
-        $bench_macro!($name5, $size5, FoldHashMap, RandomKeys::new());
-        $bench_macro!($name6, $size6, FoldHashMap, RandomKeys::new());
-        $bench_macro!($name7, $size7, FoldHashMap, RandomKeys::new());
+        $bench_macro!($name0, $size0, RapidHashMap, RandomKeys::new());
+        $bench_macro!($name1, $size1, RapidHashMap, RandomKeys::new());
+        $bench_macro!($name2, $size2, RapidHashMap, RandomKeys::new());
+        $bench_macro!($name3, $size3, RapidHashMap, RandomKeys::new());
+        $bench_macro!($name4, $size4, RapidHashMap, RandomKeys::new());
+        $bench_macro!($name5, $size5, RapidHashMap, RandomKeys::new());
+        $bench_macro!($name6, $size6, RapidHashMap, RandomKeys::new());
+        $bench_macro!($name7, $size7, RapidHashMap, RandomKeys::new());
     };
 }
 
@@ -112,11 +112,11 @@ macro_rules! bench_insert {
 
 bench_suite!(
     bench_insert,
-    insert_foldhash_serial,
+    insert_rapidhash_serial,
     insert_std_serial,
-    insert_foldhash_highbits,
+    insert_rapidhash_highbits,
     insert_std_highbits,
-    insert_foldhash_random,
+    insert_rapidhash_random,
     insert_std_random
 );
 
@@ -137,11 +137,11 @@ macro_rules! bench_grow_insert {
 
 bench_suite!(
     bench_grow_insert,
-    grow_insert_foldhash_serial,
+    grow_insert_rapidhash_serial,
     grow_insert_std_serial,
-    grow_insert_foldhash_highbits,
+    grow_insert_rapidhash_highbits,
     grow_insert_std_highbits,
-    grow_insert_foldhash_random,
+    grow_insert_rapidhash_random,
     grow_insert_std_random
 );
 
@@ -173,11 +173,11 @@ macro_rules! bench_insert_erase {
 
 bench_suite!(
     bench_insert_erase,
-    insert_erase_foldhash_serial,
+    insert_erase_rapidhash_serial,
     insert_erase_std_serial,
-    insert_erase_foldhash_highbits,
+    insert_erase_rapidhash_highbits,
     insert_erase_std_highbits,
-    insert_erase_foldhash_random,
+    insert_erase_rapidhash_random,
     insert_erase_std_random
 );
 
@@ -202,11 +202,11 @@ macro_rules! bench_lookup {
 
 bench_suite!(
     bench_lookup,
-    lookup_foldhash_serial,
+    lookup_rapidhash_serial,
     lookup_std_serial,
-    lookup_foldhash_highbits,
+    lookup_rapidhash_highbits,
     lookup_std_highbits,
-    lookup_foldhash_random,
+    lookup_rapidhash_random,
     lookup_std_random
 );
 
@@ -231,11 +231,11 @@ macro_rules! bench_lookup_fail {
 
 bench_suite!(
     bench_lookup_fail,
-    lookup_fail_foldhash_serial,
+    lookup_fail_rapidhash_serial,
     lookup_fail_std_serial,
-    lookup_fail_foldhash_highbits,
+    lookup_fail_rapidhash_highbits,
     lookup_fail_std_highbits,
-    lookup_fail_foldhash_random,
+    lookup_fail_rapidhash_random,
     lookup_fail_std_random
 );
 
@@ -336,11 +336,11 @@ macro_rules! bench_iter {
 
 bench_suite!(
     bench_iter,
-    iter_foldhash_serial,
+    iter_rapidhash_serial,
     iter_std_serial,
-    iter_foldhash_highbits,
+    iter_rapidhash_highbits,
     iter_std_highbits,
-    iter_foldhash_random,
+    iter_rapidhash_random,
     iter_std_random
 );