+++ src/map.rs
@@ -15,7 +15,7 @@ pub use crate::raw_entry::*;
 
 /// A hash map implemented with quadratic probing and SIMD lookup.
 ///
-/// The default hashing algorithm is currently [`foldhash`], though this is
+/// The default hashing algorithm is currently [`rapidhash`], though this is
 /// subject to change at any point in the future. This hash function is very
 /// fast for all types of keys, but this algorithm will typically *not* protect
 /// against attacks such as HashDoS.
@@ -142,7 +142,7 @@ pub use crate::raw_entry::*;
 /// [`with_hasher`]: #method.with_hasher
 /// [`with_capacity_and_hasher`]: #method.with_capacity_and_hasher
 /// [`fnv`]: https://crates.io/crates/fnv
-/// [`foldhash`]: https://crates.io/crates/foldhash
+/// [`rapidhash`]: https://crates.io/crates/rapidhash
 ///
 /// ```
 /// use hashbrown::HashMap;